scripts/api/requirements.txt
aniso8601==9.0.1
apispec==5.1.1
asttokens==2.0.5
backcall==0.2.0
click==8.0.4
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
Flask==2.0.3
flask-apispec==0.11.0
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
ipykernel==6.9.2
ipython==8.1.1
itsdangerous==2.1.1
jedi==0.18.1
Jinja2==3.0.3
jupyter-client==7.1.2
jupyter-core==4.9.2
MarkupSafe==2.1.1
marshmallow==3.15.0
matplotlib-inline==0.1.3
mysql-connector==2.2.9
nest-asyncio==1.5.4
numpy==1.22.3
packaging==21.3
pandas==1.4.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.28
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pyzmq==22.3.0
six==1.16.0
stack-data==0.2.0
tornado==6.1
traitlets==5.1.1
wcwidth==0.2.5
webargs==8.1.0
Werkzeug==2.0.3
